# Use RWMutex in pool