Replace `Mutex` to `RWMutex`.
Use `RLock()` to spend waiting for locks with minimal time.

Thanks!

<!-- Reviewable:start -->
---
This change is [<img src="https://reviewable.io/review_button.svg" height="34" align="absmiddle" alt="Reviewable"/>](https://reviewable.io/reviews/go-redis/redis/404)
<!-- Reviewable:end -->