@vmihailenco 2016-10-14
This makes benchmarks 20% slower

```
benchmark                          old ns/op     new ns/op     delta
BenchmarkPoolGetPut10Conns-4       360           437           +21.39%
BenchmarkPoolGetPut100Conns-4      374           456           +21.93%
BenchmarkPoolGetPut1000Conns-4     345           440           +27.54%
```

And I doubt that code like

```
	p.connsMu.Lock()
 	l := len(p.conns)
	p.connsMu.Unlock()
```

can benefit from using `RWMutex`.

If we are to micro-optimize this I would
1. use atomics to manage Len and FreeLen
2. remove synchronization from Len and FreeLen since we don't rely on these number to be super accurate

But again I doubt that will speedup real programs...