@dim 2016-10-14
I can only see this being a benefit if used in the contexts of a large,
public web app, with potentially thousands of active goroutines. In that
case, you could avoid the threads blocking each others reads, but that
doesn't really apply for the majority of use cases.