@yyoshiki41 2016-10-14
@vmihailenco @dim 

That benchmark may be correct.
In terms of implementation, `RWMutex.Lock()` is totally different from `Mutex.Lock()` :thinking: 

- [RWMutex.Lock](https://golang.org/src/sync/rwmutex.go?s=2399:2424#L70)
- [Mutex.Lock](https://golang.org/src/sync/mutex.go?s=1143:1165#L34)

This benchmark example shows same results.
(`RWMutex.Lock` is slower than `Mutex.Lock()`.)

```bash
$ go test -bench .
BenchmarkMutex-4     	   10000	    104281 ns/op
BenchmarkRWMutex-4   	   10000	    150461 ns/op
```

<details>
 <summary>Code</summary>
```go
package main

import (
	"sync"
	"testing"
)

var (
	mu    sync.Mutex
	cnt   int
	rwMu  sync.RWMutex
	rwCnt int
)

func BenchmarkMutex(b *testing.B) {
	b.ResetTimer()
	b.RunParallel(func(pb *testing.PB) {
		for pb.Next() {
			for i := 0; i < 1000; i++ {
				mu.Lock()
				cnt++
				mu.Unlock()
			}
		}
	})
}

func BenchmarkRWMutex(b *testing.B) {
	b.ResetTimer()
	b.RunParallel(func(pb *testing.PB) {
		for pb.Next() {
			for i := 0; i < 1000; i++ {
				rwMu.Lock()
				rwCnt++
				rwMu.Unlock()
			}
		}
	})
}
```
</details>

We should take the frequency of `RLcok` into consideration.
I think that the number of calling `Len()` and `FreeLen()` are small(actually zero in benchmark),
so we can't benefit from using RWMutex.

Thanks for your help!!